requirements.txt
chainlit
fastapi
uvicorn
msal
python-jose
python-dotenv
requests
